Moving useProfile hooke to src/utils/profile.tsx
